requirements.txt
datawire-quark
datawire-cloudtools
nose